# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Action-specific configuration
 */
export type AutomationActionConfig = { [key: string]: unknown };

export interface Automation {
  id: number;
  name: string;
  description?: string;
  /** GitHub event type e.g. push, pull_request, issues */
  triggerEvent: string;
  /** Optional filter condition */
  triggerCondition?: string;
  /** Action to perform e.g. label, comment, close, assign */
  actionType: string;
  /** Action-specific configuration */
  actionConfig?: AutomationActionConfig;
  enabled: boolean;
  /** owner/repo format or * for all */
  repository: string;
  createdAt: string;
  updatedAt: string;
  lastRunAt?: string | null;
  runCount: number;
}

export type CreateAutomationBodyActionConfig = { [key: string]: unknown };

export interface CreateAutomationBody {
  name: string;
  description?: string;
  triggerEvent: string;
  triggerCondition?: string;
  actionType: string;
  actionConfig?: CreateAutomationBodyActionConfig;
  enabled?: boolean;
  repository: string;
}

export type UpdateAutomationBodyActionConfig = { [key: string]: unknown };

export interface UpdateAutomationBody {
  name?: string;
  description?: string;
  triggerEvent?: string;
  triggerCondition?: string;
  actionType?: string;
  actionConfig?: UpdateAutomationBodyActionConfig;
  enabled?: boolean;
  repository?: string;
}

export type AutomationRunStatus =
  (typeof AutomationRunStatus)[keyof typeof AutomationRunStatus];

export const AutomationRunStatus = {
  success: "success",
  failure: "failure",
  skipped: "skipped",
} as const;

export type AutomationRunPayload = { [key: string]: unknown };

export interface AutomationRun {
  id: number;
  automationId: number;
  automationName: string;
  triggerEvent: string;
  repository: string;
  status: AutomationRunStatus;
  errorMessage?: string | null;
  payload?: AutomationRunPayload;
  createdAt: string;
}

export interface RunStats {
  total: number;
  success: number;
  failure: number;
  skipped: number;
  activeAutomations: number;
  totalAutomations: number;
  last24hRuns: number;
}

export type WebhookEventPayload = { [key: string]: unknown };

export interface WebhookEvent {
  id: number;
  eventType: string;
  repository: string;
  deliveryId?: string | null;
  payload: WebhookEventPayload;
  processedAt: string;
  automationsTriggered: number;
}

export interface WebhookAck {
  received: boolean;
  automationsTriggered: number;
}

export type ListRunsParams = {
  automationId?: number;
  status?: ListRunsStatus;
};

export type ListRunsStatus =
  (typeof ListRunsStatus)[keyof typeof ListRunsStatus];

export const ListRunsStatus = {
  success: "success",
  failure: "failure",
  skipped: "skipped",
} as const;

export type ReceiveGithubWebhookBody = { [key: string]: unknown };