# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: automations
    description: GitHub automation rules
  - name: webhooks
    description: GitHub webhook events
  - name: runs
    description: Automation run history

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List all automations
      description: Returns all configured GitHub automation rules
      responses:
        "200":
          description: List of automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create an automation
      description: Create a new GitHub automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"

  /automations/{id}:
    get:
      operationId: getAutomation
      tags: [automations]
      summary: Get an automation
      description: Returns a single automation rule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
        "404":
          description: Not found
    put:
      operationId: updateAutomation
      tags: [automations]
      summary: Update an automation
      description: Update an existing automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete an automation
      description: Remove an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /automations/{id}/toggle:
    post:
      operationId: toggleAutomation
      tags: [automations]
      summary: Toggle automation enabled/disabled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"

  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: List automation runs
      description: Returns recent automation execution history
      parameters:
        - name: automationId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [success, failure, skipped]
      responses:
        "200":
          description: List of runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRun"

  /runs/stats:
    get:
      operationId: getRunStats
      tags: [runs]
      summary: Get run statistics summary
      description: Returns aggregated stats about recent automation runs
      responses:
        "200":
          description: Run stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunStats"

  /webhooks/github:
    post:
      operationId: receiveGithubWebhook
      tags: [webhooks]
      summary: Receive GitHub webhook
      description: Endpoint for GitHub to send webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"

  /webhook-events:
    get:
      operationId: listWebhookEvents
      tags: [webhooks]
      summary: List recent webhook events
      description: Returns recent GitHub webhook events received
      responses:
        "200":
          description: List of webhook events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEvent"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        triggerEvent:
          type: string
          description: "GitHub event type e.g. push, pull_request, issues"
        triggerCondition:
          type: string
          description: "Optional filter condition"
        actionType:
          type: string
          description: "Action to perform e.g. label, comment, close, assign"
        actionConfig:
          type: object
          additionalProperties: true
          description: "Action-specific configuration"
        enabled:
          type: boolean
        repository:
          type: string
          description: "owner/repo format or * for all"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        runCount:
          type: integer
      required:
        - id
        - name
        - triggerEvent
        - actionType
        - enabled
        - repository
        - createdAt
        - updatedAt
        - runCount

    CreateAutomationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        triggerEvent:
          type: string
        triggerCondition:
          type: string
        actionType:
          type: string
        actionConfig:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        repository:
          type: string
      required:
        - name
        - triggerEvent
        - actionType
        - repository

    UpdateAutomationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        triggerEvent:
          type: string
        triggerCondition:
          type: string
        actionType:
          type: string
        actionConfig:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        repository:
          type: string

    AutomationRun:
      type: object
      properties:
        id:
          type: integer
        automationId:
          type: integer
        automationName:
          type: string
        triggerEvent:
          type: string
        repository:
          type: string
        status:
          type: string
          enum: [success, failure, skipped]
        errorMessage:
          type: string
          nullable: true
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - automationId
        - automationName
        - triggerEvent
        - repository
        - status
        - createdAt

    RunStats:
      type: object
      properties:
        total:
          type: integer
        success:
          type: integer
        failure:
          type: integer
        skipped:
          type: integer
        activeAutomations:
          type: integer
        totalAutomations:
          type: integer
        last24hRuns:
          type: integer
      required:
        - total
        - success
        - failure
        - skipped
        - activeAutomations
        - totalAutomations
        - last24hRuns

    WebhookEvent:
      type: object
      properties:
        id:
          type: integer
        eventType:
          type: string
        repository:
          type: string
        deliveryId:
          type: string
          nullable: true
        payload:
          type: object
          additionalProperties: true
        processedAt:
          type: string
          format: date-time
        automationsTriggered:
          type: integer
      required:
        - id
        - eventType
        - repository
        - payload
        - processedAt
        - automationsTriggered

    WebhookAck:
      type: object
      properties:
        received:
          type: boolean
        automationsTriggered:
          type: integer
      required:
        - received
        - automationsTriggered