# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all configured GitHub automation rules
 * @summary List all automations
 */
export const ListAutomationsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  triggerEvent: zod
    .string()
    .describe("GitHub event type e.g. push, pull_request, issues"),
  triggerCondition: zod
    .string()
    .optional()
    .describe("Optional filter condition"),
  actionType: zod
    .string()
    .describe("Action to perform e.g. label, comment, close, assign"),
  actionConfig: zod
    .record(zod.string(), zod.unknown())
    .optional()
    .describe("Action-specific configuration"),
  enabled: zod.boolean(),
  repository: zod.string().describe("owner\/repo format or \* for all"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastRunAt: zod.coerce.date().nullish(),
  runCount: zod.number(),
});
export const ListAutomationsResponse = zod.array(ListAutomationsResponseItem);

/**
 * Create a new GitHub automation rule
 * @summary Create an automation
 */
export const CreateAutomationBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  triggerEvent: zod.string(),
  triggerCondition: zod.string().optional(),
  actionType: zod.string(),
  actionConfig: zod.record(zod.string(), zod.unknown()).optional(),
  enabled: zod.boolean().optional(),
  repository: zod.string(),
});

/**
 * Returns a single automation rule by ID
 * @summary Get an automation
 */
export const GetAutomationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAutomationResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  triggerEvent: zod
    .string()
    .describe("GitHub event type e.g. push, pull_request, issues"),
  triggerCondition: zod
    .string()
    .optional()
    .describe("Optional filter condition"),
  actionType: zod
    .string()
    .describe("Action to perform e.g. label, comment, close, assign"),
  actionConfig: zod
    .record(zod.string(), zod.unknown())
    .optional()
    .describe("Action-specific configuration"),
  enabled: zod.boolean(),
  repository: zod.string().describe("owner\/repo format or \* for all"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastRunAt: zod.coerce.date().nullish(),
  runCount: zod.number(),
});

/**
 * Update an existing automation rule
 * @summary Update an automation
 */
export const UpdateAutomationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAutomationBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  triggerEvent: zod.string().optional(),
  triggerCondition: zod.string().optional(),
  actionType: zod.string().optional(),
  actionConfig: zod.record(zod.string(), zod.unknown()).optional(),
  enabled: zod.boolean().optional(),
  repository: zod.string().optional(),
});

export const UpdateAutomationResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  triggerEvent: zod
    .string()
    .describe("GitHub event type e.g. push, pull_request, issues"),
  triggerCondition: zod
    .string()
    .optional()
    .describe("Optional filter condition"),
  actionType: zod
    .string()
    .describe("Action to perform e.g. label, comment, close, assign"),
  actionConfig: zod
    .record(zod.string(), zod.unknown())
    .optional()
    .describe("Action-specific configuration"),
  enabled: zod.boolean(),
  repository: zod.string().describe("owner\/repo format or \* for all"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastRunAt: zod.coerce.date().nullish(),
  runCount: zod.number(),
});

/**
 * Remove an automation rule
 * @summary Delete an automation
 */
export const DeleteAutomationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle automation enabled/disabled
 */
export const ToggleAutomationParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleAutomationResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  triggerEvent: zod
    .string()
    .describe("GitHub event type e.g. push, pull_request, issues"),
  triggerCondition: zod
    .string()
    .optional()
    .describe("Optional filter condition"),
  actionType: zod
    .string()
    .describe("Action to perform e.g. label, comment, close, assign"),
  actionConfig: zod
    .record(zod.string(), zod.unknown())
    .optional()
    .describe("Action-specific configuration"),
  enabled: zod.boolean(),
  repository: zod.string().describe("owner\/repo format or \* for all"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastRunAt: zod.coerce.date().nullish(),
  runCount: zod.number(),
});

/**
 * Returns recent automation execution history
 * @summary List automation runs
 */
export const ListRunsQueryParams = zod.object({
  automationId: zod.coerce.number().optional(),
  status: zod.enum(["success", "failure", "skipped"]).optional(),
});

export const ListRunsResponseItem = zod.object({
  id: zod.number(),
  automationId: zod.number(),
  automationName: zod.string(),
  triggerEvent: zod.string(),
  repository: zod.string(),
  status: zod.enum(["success", "failure", "skipped"]),
  errorMessage: zod.string().nullish(),
  payload: zod.record(zod.string(), zod.unknown()).optional(),
  createdAt: zod.coerce.date(),
});
export const ListRunsResponse = zod.array(ListRunsResponseItem);

/**
 * Returns aggregated stats about recent automation runs
 * @summary Get run statistics summary
 */
export const GetRunStatsResponse = zod.object({
  total: zod.number(),
  success: zod.number(),
  failure: zod.number(),
  skipped: zod.number(),
  activeAutomations: zod.number(),
  totalAutomations: zod.number(),
  last24hRuns: zod.number(),
});

/**
 * Endpoint for GitHub to send webhook events
 * @summary Receive GitHub webhook
 */
export const ReceiveGithubWebhookBody = zod.record(zod.string(), zod.unknown());

export const ReceiveGithubWebhookResponse = zod.object({
  received: zod.boolean(),
  automationsTriggered: zod.number(),
});

/**
 * Returns recent GitHub webhook events received
 * @summary List recent webhook events
 */
export const ListWebhookEventsResponseItem = zod.object({
  id: zod.number(),
  eventType: zod.string(),
  repository: zod.string(),
  deliveryId: zod.string().nullish(),
  payload: zod.record(zod.string(), zod.unknown()),
  processedAt: zod.coerce.date(),
  automationsTriggered: zod.number(),
});
export const ListWebhookEventsResponse = zod.array(
  ListWebhookEventsResponseItem,
);