# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AutomationActionConfig } from "./automationActionConfig";

export interface Automation {
  id: number;
  name: string;
  description?: string;
  /** GitHub event type e.g. push, pull_request, issues */
  triggerEvent: string;
  /** Optional filter condition */
  triggerCondition?: string;
  /** Action to perform e.g. label, comment, close, assign */
  actionType: string;
  /** Action-specific configuration */
  actionConfig?: AutomationActionConfig;
  enabled: boolean;
  /** owner/repo format or * for all */
  repository: string;
  createdAt: Date;
  updatedAt: Date;
  lastRunAt?: Date | null;
  runCount: number;
}