# lib/api-zod/src/generated/types/automationRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AutomationRunPayload } from "./automationRunPayload";
import type { AutomationRunStatus } from "./automationRunStatus";

export interface AutomationRun {
  id: number;
  automationId: number;
  automationName: string;
  triggerEvent: string;
  repository: string;
  status: AutomationRunStatus;
  errorMessage?: string | null;
  payload?: AutomationRunPayload;
  createdAt: Date;
}