# lib/api-zod/src/generated/types/createAutomationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAutomationBodyActionConfig } from "./createAutomationBodyActionConfig";

export interface CreateAutomationBody {
  name: string;
  description?: string;
  triggerEvent: string;
  triggerCondition?: string;
  actionType: string;
  actionConfig?: CreateAutomationBodyActionConfig;
  enabled?: boolean;
  repository: string;
}