# lib/api-zod/src/generated/types/runStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RunStats {
  total: number;
  success: number;
  failure: number;
  skipped: number;
  activeAutomations: number;
  totalAutomations: number;
  last24hRuns: number;
}