# lib/api-zod/src/generated/types/updateAutomationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAutomationBodyActionConfig } from "./updateAutomationBodyActionConfig";

export interface UpdateAutomationBody {
  name?: string;
  description?: string;
  triggerEvent?: string;
  triggerCondition?: string;
  actionType?: string;
  actionConfig?: UpdateAutomationBodyActionConfig;
  enabled?: boolean;
  repository?: string;
}