# lib/api-zod/src/generated/types/webhookEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WebhookEventPayload } from "./webhookEventPayload";

export interface WebhookEvent {
  id: number;
  eventType: string;
  repository: string;
  deliveryId?: string | null;
  payload: WebhookEventPayload;
  processedAt: Date;
  automationsTriggered: number;
}